requirements.txt
# Core AI and Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0

# Web Framework and UI
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
jinja2>=3.1.0

# Data Processing and Analytics
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and Logging
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Database and Storage
sqlalchemy>=2.0.0

# Real-time Communication
websockets>=11.0.0

# Image Processing
imageio>=2.31.0
scikit-image>=0.21.0

# Google Drive Integration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Utilities
requests>=2.31.0
psutil>=5.9.0
schedule>=1.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0 